+++ services/envelope-store/store.pb.go
@@ -7,9 +7,11 @@ import (
 	context "context"
 	fmt "fmt"
 	proto "github.com/golang/protobuf/proto"
+	duration "github.com/golang/protobuf/ptypes/duration"
 	timestamp "github.com/golang/protobuf/ptypes/timestamp"
 	chain "gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/types/chain"
 	envelope "gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/types/envelope"
+	ethereum "gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/types/ethereum"
 	grpc "google.golang.org/grpc"
 	math "math"
 )
@@ -25,6 +27,111 @@ var _ = math.Inf
 // proto package needs to be updated.
 const _ = proto.ProtoPackageIsVersion3 // please upgrade the proto package
 
+// Possible statuses for a transaction
+type Status int32
+
+const (
+	Status_STORED  Status = 0
+	Status_PENDING Status = 1
+	Status_MINED   Status = 2
+	Status_ERROR   Status = 3
+)
+
+var Status_name = map[int32]string{
+	0: "STORED",
+	1: "PENDING",
+	2: "MINED",
+	3: "ERROR",
+}
+
+var Status_value = map[string]int32{
+	"STORED":  0,
+	"PENDING": 1,
+	"MINED":   2,
+	"ERROR":   3,
+}
+
+func (x Status) String() string {
+	return proto.EnumName(Status_name, int32(x))
+}
+
+func (Status) EnumDescriptor() ([]byte, []int) {
+	return fileDescriptor_3251193be0eb6bb8, []int{0}
+}
+
+type StatusInfo struct {
+	// Status of Transaction
+	Status Status `protobuf:"varint,1,opt,name=status,proto3,enum=envelopestore.Status" json:"status,omitempty"`
+	// Date events for envelope
+	StoredAt             *timestamp.Timestamp `protobuf:"bytes,2,opt,name=storedAt,proto3" json:"storedAt,omitempty"`
+	SentAt               *timestamp.Timestamp `protobuf:"bytes,3,opt,name=sentAt,proto3" json:"sentAt,omitempty"`
+	MinedAt              *timestamp.Timestamp `protobuf:"bytes,4,opt,name=minedAt,proto3" json:"minedAt,omitempty"`
+	ErrorAt              *timestamp.Timestamp `protobuf:"bytes,5,opt,name=errorAt,proto3" json:"errorAt,omitempty"`
+	XXX_NoUnkeyedLiteral struct{}             `json:"-"`
+	XXX_unrecognized     []byte               `json:"-"`
+	XXX_sizecache        int32                `json:"-"`
+}
+
+func (m *StatusInfo) Reset()         { *m = StatusInfo{} }
+func (m *StatusInfo) String() string { return proto.CompactTextString(m) }
+func (*StatusInfo) ProtoMessage()    {}
+func (*StatusInfo) Descriptor() ([]byte, []int) {
+	return fileDescriptor_3251193be0eb6bb8, []int{0}
+}
+
+func (m *StatusInfo) XXX_Unmarshal(b []byte) error {
+	return xxx_messageInfo_StatusInfo.Unmarshal(m, b)
+}
+func (m *StatusInfo) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	return xxx_messageInfo_StatusInfo.Marshal(b, m, deterministic)
+}
+func (m *StatusInfo) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_StatusInfo.Merge(m, src)
+}
+func (m *StatusInfo) XXX_Size() int {
+	return xxx_messageInfo_StatusInfo.Size(m)
+}
+func (m *StatusInfo) XXX_DiscardUnknown() {
+	xxx_messageInfo_StatusInfo.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_StatusInfo proto.InternalMessageInfo
+
+func (m *StatusInfo) GetStatus() Status {
+	if m != nil {
+		return m.Status
+	}
+	return Status_STORED
+}
+
+func (m *StatusInfo) GetStoredAt() *timestamp.Timestamp {
+	if m != nil {
+		return m.StoredAt
+	}
+	return nil
+}
+
+func (m *StatusInfo) GetSentAt() *timestamp.Timestamp {
+	if m != nil {
+		return m.SentAt
+	}
+	return nil
+}
+
+func (m *StatusInfo) GetMinedAt() *timestamp.Timestamp {
+	if m != nil {
+		return m.MinedAt
+	}
+	return nil
+}
+
+func (m *StatusInfo) GetErrorAt() *timestamp.Timestamp {
+	if m != nil {
+		return m.ErrorAt
+	}
+	return nil
+}
+
 type StoreRequest struct {
 	Envelope             *envelope.Envelope `protobuf:"bytes,1,opt,name=envelope,proto3" json:"envelope,omitempty"`
 	XXX_NoUnkeyedLiteral struct{}           `json:"-"`
@@ -36,7 +143,7 @@ func (m *StoreRequest) Reset()         { *m = StoreRequest{} }
 func (m *StoreRequest) String() string { return proto.CompactTextString(m) }
 func (*StoreRequest) ProtoMessage()    {}
 func (*StoreRequest) Descriptor() ([]byte, []int) {
-	return fileDescriptor_3251193be0eb6bb8, []int{0}
+	return fileDescriptor_3251193be0eb6bb8, []int{1}
 }
 
 func (m *StoreRequest) XXX_Unmarshal(b []byte) error {
@@ -65,12 +172,9 @@ func (m *StoreRequest) GetEnvelope() *envelope.Envelope {
 }
 
 type StoreResponse struct {
-	// Status of Envelope
-	Status string `protobuf:"bytes,1,opt,name=status,proto3" json:"status,omitempty"`
-	// Last update date of envelope stored
-	LastUpdated *timestamp.Timestamp `protobuf:"bytes,2,opt,name=last_updated,json=lastUpdated,proto3" json:"last_updated,omitempty"`
 	// Envelope
-	Envelope             *envelope.Envelope `protobuf:"bytes,3,opt,name=envelope,proto3" json:"envelope,omitempty"`
+	Envelope             *envelope.Envelope `protobuf:"bytes,1,opt,name=envelope,proto3" json:"envelope,omitempty"`
+	StatusInfo           *StatusInfo        `protobuf:"bytes,2,opt,name=status_info,json=statusInfo,proto3" json:"status_info,omitempty"`
 	XXX_NoUnkeyedLiteral struct{}           `json:"-"`
 	XXX_unrecognized     []byte             `json:"-"`
 	XXX_sizecache        int32              `json:"-"`
@@ -80,7 +184,7 @@ func (m *StoreResponse) Reset()         { *m = StoreResponse{} }
 func (m *StoreResponse) String() string { return proto.CompactTextString(m) }
 func (*StoreResponse) ProtoMessage()    {}
 func (*StoreResponse) Descriptor() ([]byte, []int) {
-	return fileDescriptor_3251193be0eb6bb8, []int{1}
+	return fileDescriptor_3251193be0eb6bb8, []int{2}
 }
 
 func (m *StoreResponse) XXX_Unmarshal(b []byte) error {
@@ -101,20 +205,6 @@ func (m *StoreResponse) XXX_DiscardUnknown() {
 
 var xxx_messageInfo_StoreResponse proto.InternalMessageInfo
 
-func (m *StoreResponse) GetStatus() string {
-	if m != nil {
-		return m.Status
-	}
-	return ""
-}
-
-func (m *StoreResponse) GetLastUpdated() *timestamp.Timestamp {
-	if m != nil {
-		return m.LastUpdated
-	}
-	return nil
-}
-
 func (m *StoreResponse) GetEnvelope() *envelope.Envelope {
 	if m != nil {
 		return m.Envelope
@@ -122,56 +212,14 @@ func (m *StoreResponse) GetEnvelope() *envelope.Envelope {
 	return nil
 }
 
-type TxHashRequest struct {
-	// Chain ID the transaction has been sent to
-	ChainId *chain.Chain `protobuf:"bytes,1,opt,name=chain_id,json=chainId,proto3" json:"chain_id,omitempty"`
-	// Hash of the transaction
-	TxHash               string   `protobuf:"bytes,2,opt,name=tx_hash,json=txHash,proto3" json:"tx_hash,omitempty"`
-	XXX_NoUnkeyedLiteral struct{} `json:"-"`
-	XXX_unrecognized     []byte   `json:"-"`
-	XXX_sizecache        int32    `json:"-"`
-}
-
-func (m *TxHashRequest) Reset()         { *m = TxHashRequest{} }
-func (m *TxHashRequest) String() string { return proto.CompactTextString(m) }
-func (*TxHashRequest) ProtoMessage()    {}
-func (*TxHashRequest) Descriptor() ([]byte, []int) {
-	return fileDescriptor_3251193be0eb6bb8, []int{2}
-}
-
-func (m *TxHashRequest) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_TxHashRequest.Unmarshal(m, b)
-}
-func (m *TxHashRequest) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_TxHashRequest.Marshal(b, m, deterministic)
-}
-func (m *TxHashRequest) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_TxHashRequest.Merge(m, src)
-}
-func (m *TxHashRequest) XXX_Size() int {
-	return xxx_messageInfo_TxHashRequest.Size(m)
-}
-func (m *TxHashRequest) XXX_DiscardUnknown() {
-	xxx_messageInfo_TxHashRequest.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_TxHashRequest proto.InternalMessageInfo
-
-func (m *TxHashRequest) GetChainId() *chain.Chain {
+func (m *StoreResponse) GetStatusInfo() *StatusInfo {
 	if m != nil {
-		return m.ChainId
+		return m.StatusInfo
 	}
 	return nil
 }
 
-func (m *TxHashRequest) GetTxHash() string {
-	if m != nil {
-		return m.TxHash
-	}
-	return ""
-}
-
-type IDRequest struct {
+type LoadByIDRequest struct {
 	// Envelope identifier
 	Id                   string   `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
 	XXX_NoUnkeyedLiteral struct{} `json:"-"`
@@ -179,43 +227,92 @@ type IDRequest struct {
 	XXX_sizecache        int32    `json:"-"`
 }
 
-func (m *IDRequest) Reset()         { *m = IDRequest{} }
-func (m *IDRequest) String() string { return proto.CompactTextString(m) }
-func (*IDRequest) ProtoMessage()    {}
-func (*IDRequest) Descriptor() ([]byte, []int) {
+func (m *LoadByIDRequest) Reset()         { *m = LoadByIDRequest{} }
+func (m *LoadByIDRequest) String() string { return proto.CompactTextString(m) }
+func (*LoadByIDRequest) ProtoMessage()    {}
+func (*LoadByIDRequest) Descriptor() ([]byte, []int) {
 	return fileDescriptor_3251193be0eb6bb8, []int{3}
 }
 
-func (m *IDRequest) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_IDRequest.Unmarshal(m, b)
+func (m *LoadByIDRequest) XXX_Unmarshal(b []byte) error {
+	return xxx_messageInfo_LoadByIDRequest.Unmarshal(m, b)
 }
-func (m *IDRequest) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_IDRequest.Marshal(b, m, deterministic)
+func (m *LoadByIDRequest) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	return xxx_messageInfo_LoadByIDRequest.Marshal(b, m, deterministic)
 }
-func (m *IDRequest) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_IDRequest.Merge(m, src)
+func (m *LoadByIDRequest) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_LoadByIDRequest.Merge(m, src)
 }
-func (m *IDRequest) XXX_Size() int {
-	return xxx_messageInfo_IDRequest.Size(m)
+func (m *LoadByIDRequest) XXX_Size() int {
+	return xxx_messageInfo_LoadByIDRequest.Size(m)
 }
-func (m *IDRequest) XXX_DiscardUnknown() {
-	xxx_messageInfo_IDRequest.DiscardUnknown(m)
+func (m *LoadByIDRequest) XXX_DiscardUnknown() {
+	xxx_messageInfo_LoadByIDRequest.DiscardUnknown(m)
 }
 
-var xxx_messageInfo_IDRequest proto.InternalMessageInfo
+var xxx_messageInfo_LoadByIDRequest proto.InternalMessageInfo
 
-func (m *IDRequest) GetId() string {
+func (m *LoadByIDRequest) GetId() string {
 	if m != nil {
 		return m.Id
 	}
 	return ""
 }
 
+type LoadByTxHashRequest struct {
+	// Chain the transaction has been sent to
+	Chain *chain.Chain `protobuf:"bytes,2,opt,name=chain,proto3" json:"chain,omitempty"`
+	// Hash of the transaction
+	TxHash               *ethereum.Hash `protobuf:"bytes,3,opt,name=tx_hash,json=txHash,proto3" json:"tx_hash,omitempty"`
+	XXX_NoUnkeyedLiteral struct{}       `json:"-"`
+	XXX_unrecognized     []byte         `json:"-"`
+	XXX_sizecache        int32          `json:"-"`
+}
+
+func (m *LoadByTxHashRequest) Reset()         { *m = LoadByTxHashRequest{} }
+func (m *LoadByTxHashRequest) String() string { return proto.CompactTextString(m) }
+func (*LoadByTxHashRequest) ProtoMessage()    {}
+func (*LoadByTxHashRequest) Descriptor() ([]byte, []int) {
+	return fileDescriptor_3251193be0eb6bb8, []int{4}
+}
+
+func (m *LoadByTxHashRequest) XXX_Unmarshal(b []byte) error {
+	return xxx_messageInfo_LoadByTxHashRequest.Unmarshal(m, b)
+}
+func (m *LoadByTxHashRequest) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	return xxx_messageInfo_LoadByTxHashRequest.Marshal(b, m, deterministic)
+}
+func (m *LoadByTxHashRequest) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_LoadByTxHashRequest.Merge(m, src)
+}
+func (m *LoadByTxHashRequest) XXX_Size() int {
+	return xxx_messageInfo_LoadByTxHashRequest.Size(m)
+}
+func (m *LoadByTxHashRequest) XXX_DiscardUnknown() {
+	xxx_messageInfo_LoadByTxHashRequest.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_LoadByTxHashRequest proto.InternalMessageInfo
+
+func (m *LoadByTxHashRequest) GetChain() *chain.Chain {
+	if m != nil {
+		return m.Chain
+	}
+	return nil
+}
+
+func (m *LoadByTxHashRequest) GetTxHash() *ethereum.Hash {
+	if m != nil {
+		return m.TxHash
+	}
+	return nil
+}
+
 type SetStatusRequest struct {
 	// Envelope identifier
 	Id string `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
 	// Status
-	Status               string   `protobuf:"bytes,2,opt,name=status,proto3" json:"status,omitempty"`
+	Status               Status   `protobuf:"varint,2,opt,name=status,proto3,enum=envelopestore.Status" json:"status,omitempty"`
 	XXX_NoUnkeyedLiteral struct{} `json:"-"`
 	XXX_unrecognized     []byte   `json:"-"`
 	XXX_sizecache        int32    `json:"-"`
@@ -225,7 +322,7 @@ func (m *SetStatusRequest) Reset()         { *m = SetStatusRequest{} }
 func (m *SetStatusRequest) String() string { return proto.CompactTextString(m) }
 func (*SetStatusRequest) ProtoMessage()    {}
 func (*SetStatusRequest) Descriptor() ([]byte, []int) {
-	return fileDescriptor_3251193be0eb6bb8, []int{4}
+	return fileDescriptor_3251193be0eb6bb8, []int{5}
 }
 
 func (m *SetStatusRequest) XXX_Unmarshal(b []byte) error {
@@ -253,57 +350,65 @@ func (m *SetStatusRequest) GetId() string {
 	return ""
 }
 
-func (m *SetStatusRequest) GetStatus() string {
+func (m *SetStatusRequest) GetStatus() Status {
 	if m != nil {
 		return m.Status
 	}
-	return ""
+	return Status_STORED
 }
 
-type SetStatusResponse struct {
-	XXX_NoUnkeyedLiteral struct{} `json:"-"`
-	XXX_unrecognized     []byte   `json:"-"`
-	XXX_sizecache        int32    `json:"-"`
+type StatusResponse struct {
+	StatusInfo           *StatusInfo `protobuf:"bytes,1,opt,name=status_info,json=statusInfo,proto3" json:"status_info,omitempty"`
+	XXX_NoUnkeyedLiteral struct{}    `json:"-"`
+	XXX_unrecognized     []byte      `json:"-"`
+	XXX_sizecache        int32       `json:"-"`
 }
 
-func (m *SetStatusResponse) Reset()         { *m = SetStatusResponse{} }
-func (m *SetStatusResponse) String() string { return proto.CompactTextString(m) }
-func (*SetStatusResponse) ProtoMessage()    {}
-func (*SetStatusResponse) Descriptor() ([]byte, []int) {
-	return fileDescriptor_3251193be0eb6bb8, []int{5}
+func (m *StatusResponse) Reset()         { *m = StatusResponse{} }
+func (m *StatusResponse) String() string { return proto.CompactTextString(m) }
+func (*StatusResponse) ProtoMessage()    {}
+func (*StatusResponse) Descriptor() ([]byte, []int) {
+	return fileDescriptor_3251193be0eb6bb8, []int{6}
 }
 
-func (m *SetStatusResponse) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_SetStatusResponse.Unmarshal(m, b)
+func (m *StatusResponse) XXX_Unmarshal(b []byte) error {
+	return xxx_messageInfo_StatusResponse.Unmarshal(m, b)
 }
-func (m *SetStatusResponse) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_SetStatusResponse.Marshal(b, m, deterministic)
+func (m *StatusResponse) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	return xxx_messageInfo_StatusResponse.Marshal(b, m, deterministic)
 }
-func (m *SetStatusResponse) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_SetStatusResponse.Merge(m, src)
+func (m *StatusResponse) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_StatusResponse.Merge(m, src)
 }
-func (m *SetStatusResponse) XXX_Size() int {
-	return xxx_messageInfo_SetStatusResponse.Size(m)
+func (m *StatusResponse) XXX_Size() int {
+	return xxx_messageInfo_StatusResponse.Size(m)
 }
-func (m *SetStatusResponse) XXX_DiscardUnknown() {
-	xxx_messageInfo_SetStatusResponse.DiscardUnknown(m)
+func (m *StatusResponse) XXX_DiscardUnknown() {
+	xxx_messageInfo_StatusResponse.DiscardUnknown(m)
 }
 
-var xxx_messageInfo_SetStatusResponse proto.InternalMessageInfo
+var xxx_messageInfo_StatusResponse proto.InternalMessageInfo
+
+func (m *StatusResponse) GetStatusInfo() *StatusInfo {
+	if m != nil {
+		return m.StatusInfo
+	}
+	return nil
+}
 
 type LoadPendingRequest struct {
-	// Pending duration in nanoseconds
-	Duration             int64    `protobuf:"varint,1,opt,name=duration,proto3" json:"duration,omitempty"`
-	XXX_NoUnkeyedLiteral struct{} `json:"-"`
-	XXX_unrecognized     []byte   `json:"-"`
-	XXX_sizecache        int32    `json:"-"`
+	// Pending duration
+	Duration             *duration.Duration `protobuf:"bytes,1,opt,name=duration,proto3" json:"duration,omitempty"`
+	XXX_NoUnkeyedLiteral struct{}           `json:"-"`
+	XXX_unrecognized     []byte             `json:"-"`
+	XXX_sizecache        int32              `json:"-"`
 }
 
 func (m *LoadPendingRequest) Reset()         { *m = LoadPendingRequest{} }
 func (m *LoadPendingRequest) String() string { return proto.CompactTextString(m) }
 func (*LoadPendingRequest) ProtoMessage()    {}
 func (*LoadPendingRequest) Descriptor() ([]byte, []int) {
-	return fileDescriptor_3251193be0eb6bb8, []int{6}
+	return fileDescriptor_3251193be0eb6bb8, []int{7}
 }
 
 func (m *LoadPendingRequest) XXX_Unmarshal(b []byte) error {
@@ -324,26 +429,26 @@ func (m *LoadPendingRequest) XXX_DiscardUnknown() {
 
 var xxx_messageInfo_LoadPendingRequest proto.InternalMessageInfo
 
-func (m *LoadPendingRequest) GetDuration() int64 {
+func (m *LoadPendingRequest) GetDuration() *duration.Duration {
 	if m != nil {
 		return m.Duration
 	}
-	return 0
+	return nil
 }
 
 type LoadPendingResponse struct {
 	// Pending envelopes
-	Envelopes            []*envelope.Envelope `protobuf:"bytes,1,rep,name=envelopes,proto3" json:"envelopes,omitempty"`
-	XXX_NoUnkeyedLiteral struct{}             `json:"-"`
-	XXX_unrecognized     []byte               `json:"-"`
-	XXX_sizecache        int32                `json:"-"`
+	Responses            []*StoreResponse `protobuf:"bytes,1,rep,name=responses,proto3" json:"responses,omitempty"`
+	XXX_NoUnkeyedLiteral struct{}         `json:"-"`
+	XXX_unrecognized     []byte           `json:"-"`
+	XXX_sizecache        int32            `json:"-"`
 }
 
 func (m *LoadPendingResponse) Reset()         { *m = LoadPendingResponse{} }
 func (m *LoadPendingResponse) String() string { return proto.CompactTextString(m) }
 func (*LoadPendingResponse) ProtoMessage()    {}
 func (*LoadPendingResponse) Descriptor() ([]byte, []int) {
-	return fileDescriptor_3251193be0eb6bb8, []int{7}
+	return fileDescriptor_3251193be0eb6bb8, []int{8}
 }
 
 func (m *LoadPendingResponse) XXX_Unmarshal(b []byte) error {
@@ -364,20 +469,22 @@ func (m *LoadPendingResponse) XXX_DiscardUnknown() {
 
 var xxx_messageInfo_LoadPendingResponse proto.InternalMessageInfo
 
-func (m *LoadPendingResponse) GetEnvelopes() []*envelope.Envelope {
+func (m *LoadPendingResponse) GetResponses() []*StoreResponse {
 	if m != nil {
-		return m.Envelopes
+		return m.Responses
 	}
 	return nil
 }
 
 func init() {
+	proto.RegisterEnum("envelopestore.Status", Status_name, Status_value)
+	proto.RegisterType((*StatusInfo)(nil), "envelopestore.StatusInfo")
 	proto.RegisterType((*StoreRequest)(nil), "envelopestore.StoreRequest")
 	proto.RegisterType((*StoreResponse)(nil), "envelopestore.StoreResponse")
-	proto.RegisterType((*TxHashRequest)(nil), "envelopestore.TxHashRequest")
-	proto.RegisterType((*IDRequest)(nil), "envelopestore.IDRequest")
+	proto.RegisterType((*LoadByIDRequest)(nil), "envelopestore.LoadByIDRequest")
+	proto.RegisterType((*LoadByTxHashRequest)(nil), "envelopestore.LoadByTxHashRequest")
 	proto.RegisterType((*SetStatusRequest)(nil), "envelopestore.SetStatusRequest")
-	proto.RegisterType((*SetStatusResponse)(nil), "envelopestore.SetStatusResponse")
+	proto.RegisterType((*StatusResponse)(nil), "envelopestore.StatusResponse")
 	proto.RegisterType((*LoadPendingRequest)(nil), "envelopestore.LoadPendingRequest")
 	proto.RegisterType((*LoadPendingResponse)(nil), "envelopestore.LoadPendingResponse")
 }
@@ -387,39 +494,48 @@ func init() {
 }
 
 var fileDescriptor_3251193be0eb6bb8 = []byte{
-	// 512 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x9c, 0x94, 0x51, 0x6f, 0xd3, 0x30,
-	0x10, 0xc7, 0xd5, 0x56, 0x74, 0xcd, 0xb5, 0x45, 0xe0, 0x49, 0xac, 0xca, 0x40, 0x2b, 0xe1, 0x01,
-	0x5e, 0xb0, 0xa7, 0xf1, 0x86, 0x04, 0x0f, 0xed, 0x60, 0x2b, 0x42, 0x08, 0xd2, 0xf2, 0xc2, 0x4b,
-	0xe5, 0x26, 0x5e, 0x6a, 0x2d, 0x8d, 0x43, 0xec, 0x4c, 0xeb, 0x17, 0xe1, 0x23, 0xf1, 0xb9, 0x50,
-	0xec, 0x24, 0x24, 0x81, 0x52, 0xb4, 0x17, 0xcb, 0x97, 0xfb, 0xdf, 0xcf, 0xe7, 0xbb, 0x8b, 0xe1,
-	0x99, 0x64, 0xc9, 0x0d, 0xf7, 0x98, 0x24, 0x2c, 0xba, 0x61, 0xa1, 0x88, 0xd9, 0x4b, 0xa9, 0x44,
-	0xc2, 0x88, 0x5e, 0x71, 0x9c, 0x08, 0x25, 0xd0, 0xb0, 0xf0, 0xe9, 0x8f, 0xf6, 0x13, 0xb5, 0x8d,
-	0x2b, 0x01, 0xe5, 0xc6, 0xa8, 0xed, 0x23, 0xe3, 0xf6, 0xd6, 0x94, 0x47, 0x66, 0xcd, 0x1d, 0x27,
-	0x81, 0x10, 0x41, 0xc8, 0x88, 0xb6, 0x56, 0xe9, 0x15, 0x51, 0x7c, 0xc3, 0xa4, 0xa2, 0x9b, 0xd8,
-	0x08, 0x9c, 0xb7, 0x30, 0x98, 0x67, 0x27, 0xb8, 0xec, 0x7b, 0xca, 0xa4, 0x42, 0x18, 0x7a, 0x05,
-	0x7b, 0xd4, 0x1a, 0xb7, 0x5e, 0xf4, 0xcf, 0x10, 0x2e, 0x0f, 0x7b, 0x97, 0x6f, 0xdc, 0x52, 0xe3,
-	0xfc, 0x68, 0xc1, 0x30, 0x07, 0xc8, 0x58, 0x44, 0x92, 0xa1, 0x47, 0xd0, 0x95, 0x8a, 0xaa, 0x54,
-	0xea, 0x78, 0xcb, 0xcd, 0x2d, 0xf4, 0x06, 0x06, 0x21, 0x95, 0x6a, 0x99, 0xc6, 0x3e, 0x55, 0xcc,
-	0x1f, 0xb5, 0x35, 0xdd, 0xc6, 0x26, 0x43, 0x5c, 0x64, 0x88, 0x17, 0x45, 0x86, 0x6e, 0x3f, 0xd3,
-	0x7f, 0x35, 0xf2, 0x5a, 0x62, 0x9d, 0xff, 0x48, 0xec, 0x0b, 0x0c, 0x17, 0xb7, 0x97, 0x54, 0xae,
-	0x8b, 0x9b, 0x3d, 0x87, 0x9e, 0xae, 0xcc, 0x92, 0xfb, 0xf9, 0xcd, 0x06, 0xd8, 0x94, 0x6a, 0x9a,
-	0xad, 0xee, 0x81, 0x36, 0x66, 0x3e, 0x3a, 0x82, 0x03, 0x75, 0xbb, 0x5c, 0x53, 0xb9, 0xd6, 0x39,
-	0x5a, 0x6e, 0x57, 0x69, 0x90, 0x73, 0x0c, 0xd6, 0xec, 0xbc, 0xc0, 0xdd, 0x87, 0x76, 0x0e, 0xb2,
-	0xdc, 0x36, 0xf7, 0x9d, 0xd7, 0xf0, 0x60, 0xce, 0xd4, 0x5c, 0xdf, 0x75, 0x87, 0xa6, 0x52, 0x9a,
-	0x76, 0xb5, 0x34, 0xce, 0x21, 0x3c, 0xac, 0xc4, 0x9a, 0x3a, 0x3a, 0xa7, 0x80, 0x3e, 0x0a, 0xea,
-	0x7f, 0x66, 0x91, 0xcf, 0xa3, 0xa0, 0x40, 0xda, 0xd0, 0xf3, 0xd3, 0x84, 0x2a, 0x2e, 0x22, 0x0d,
-	0xee, 0xb8, 0xa5, 0xed, 0x5c, 0xc0, 0x61, 0x2d, 0x22, 0x6f, 0xc8, 0x29, 0x58, 0xe5, 0x30, 0x8d,
-	0x5a, 0xe3, 0xce, 0x8e, 0xd2, 0xfd, 0x16, 0x9d, 0xfd, 0xec, 0xc0, 0x3d, 0xdd, 0x54, 0x34, 0x29,
-	0x36, 0xc7, 0xb8, 0x36, 0x90, 0xb8, 0x3a, 0x34, 0xf6, 0xe3, 0xbf, 0x3b, 0xf3, 0xf3, 0x3f, 0xc0,
-	0x20, 0x4b, 0x6b, 0xb2, 0x35, 0xfd, 0x40, 0x4d, 0x75, 0xad, 0x4d, 0x7b, 0x58, 0x13, 0xe8, 0x19,
-	0xd6, 0xec, 0x1c, 0x8d, 0x1a, 0xca, 0xb2, 0x37, 0x7b, 0x18, 0x53, 0xb0, 0x2e, 0x8a, 0x6a, 0xdf,
-	0x19, 0xf2, 0x09, 0xac, 0xb2, 0x65, 0xe8, 0xa4, 0x29, 0x6d, 0x0c, 0x82, 0x3d, 0xde, 0x2d, 0xc8,
-	0x79, 0x0b, 0xe8, 0x57, 0x7a, 0x87, 0x9e, 0x36, 0x02, 0xfe, 0x9c, 0x04, 0xdb, 0xf9, 0x97, 0xc4,
-	0x50, 0x27, 0x97, 0xdf, 0xde, 0x07, 0x5c, 0x85, 0x74, 0x85, 0x3d, 0xb1, 0x21, 0xd3, 0xec, 0x5b,
-	0x34, 0xdf, 0x4a, 0xe2, 0x85, 0x9c, 0x45, 0x8a, 0x5c, 0x25, 0xc4, 0x13, 0x49, 0xf6, 0xfc, 0x50,
-	0xef, 0x9a, 0xc4, 0xd7, 0x01, 0x0e, 0xb8, 0x22, 0xf5, 0xa7, 0xc6, 0xbc, 0x4d, 0xab, 0xae, 0xfe,
-	0x3f, 0x5f, 0xfd, 0x0a, 0x00, 0x00, 0xff, 0xff, 0x3a, 0xd7, 0x71, 0x2b, 0xbd, 0x04, 0x00, 0x00,
+	// 651 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x94, 0x55, 0x5b, 0x6f, 0x12, 0x41,
+	0x14, 0x16, 0x10, 0x0a, 0x87, 0x16, 0xc9, 0x18, 0x23, 0x45, 0x7b, 0x5b, 0x1f, 0x6c, 0x4c, 0x3a,
+	0x9b, 0xe0, 0x2d, 0xf1, 0xc1, 0xa4, 0x2d, 0x68, 0x49, 0x6b, 0x2f, 0x03, 0x4f, 0xbe, 0x34, 0x0b,
+	0x0c, 0x30, 0x29, 0xec, 0xac, 0x3b, 0x43, 0x53, 0xe2, 0xbf, 0xd0, 0x3f, 0x6c, 0x76, 0x2e, 0x2b,
+	0xac, 0x9b, 0x62, 0x5f, 0x26, 0x67, 0xe7, 0x7c, 0xe7, 0x3b, 0x67, 0xbf, 0xf9, 0x26, 0x03, 0xaf,
+	0x04, 0x0d, 0x6f, 0x59, 0x9f, 0x0a, 0x97, 0xfa, 0xb7, 0x74, 0xc2, 0x03, 0x7a, 0x20, 0x24, 0x0f,
+	0xa9, 0xab, 0x56, 0x1c, 0x84, 0x5c, 0x72, 0xb4, 0x61, 0x73, 0x6a, 0xb3, 0xbe, 0x25, 0xe7, 0xc1,
+	0x42, 0x41, 0x1c, 0x68, 0x74, 0x7d, 0xd3, 0xa4, 0xe5, 0x98, 0x86, 0x74, 0x36, 0x75, 0x7b, 0x9e,
+	0xb0, 0xa9, 0xe7, 0x3a, 0xd5, 0x1f, 0x7b, 0xcc, 0xd7, 0xab, 0x49, 0xec, 0x8c, 0x38, 0x1f, 0x4d,
+	0xa8, 0xab, 0xbe, 0x7a, 0xb3, 0xa1, 0x2b, 0xd9, 0x94, 0x0a, 0xe9, 0x4d, 0x03, 0x03, 0xd8, 0x4e,
+	0x02, 0x06, 0xb3, 0xd0, 0x93, 0x8c, 0x1b, 0x02, 0xe7, 0x77, 0x16, 0xa0, 0x23, 0x3d, 0x39, 0x13,
+	0x6d, 0x7f, 0xc8, 0xd1, 0x01, 0x14, 0x84, 0xfa, 0xaa, 0x65, 0x76, 0x33, 0xfb, 0x95, 0xc6, 0x33,
+	0xbc, 0xf4, 0x0b, 0x58, 0x43, 0x89, 0x01, 0xa1, 0x0f, 0x50, 0x54, 0xfb, 0x83, 0x43, 0x59, 0xcb,
+	0xee, 0x66, 0xf6, 0xcb, 0x8d, 0x3a, 0xd6, 0x0d, 0xb1, 0x6d, 0x88, 0xbb, 0x76, 0x22, 0x12, 0x63,
+	0x51, 0x03, 0x0a, 0x82, 0xfa, 0xf2, 0x50, 0xd6, 0x72, 0x2b, 0xab, 0x0c, 0x12, 0xbd, 0x83, 0xb5,
+	0x29, 0xf3, 0x55, 0xab, 0xc7, 0x2b, 0x8b, 0x2c, 0x34, 0xaa, 0xa2, 0x61, 0xc8, 0xc3, 0x43, 0x59,
+	0xcb, 0xaf, 0xae, 0x32, 0x50, 0xe7, 0x33, 0xac, 0x77, 0xa2, 0x59, 0x09, 0xfd, 0x31, 0xa3, 0x42,
+	0x22, 0x0c, 0x45, 0xab, 0x83, 0x12, 0xa6, 0xdc, 0x40, 0xb1, 0x30, 0xb8, 0x65, 0x02, 0x12, 0x63,
+	0x9c, 0x9f, 0xb0, 0x61, 0xea, 0x45, 0xc0, 0x7d, 0x41, 0x1f, 0x4a, 0x80, 0x3e, 0x41, 0x59, 0x4b,
+	0x7c, 0xcd, 0xfc, 0x21, 0x37, 0xda, 0x6e, 0xa6, 0x1e, 0x46, 0x74, 0x6e, 0x04, 0x44, 0x1c, 0x3b,
+	0x7b, 0xf0, 0xe4, 0x8c, 0x7b, 0x83, 0xa3, 0x79, 0xbb, 0x69, 0xe7, 0xaf, 0x40, 0x96, 0x0d, 0x54,
+	0xe3, 0x12, 0xc9, 0xb2, 0x81, 0xd3, 0x83, 0xa7, 0x1a, 0xd2, 0xbd, 0x3b, 0xf1, 0xc4, 0xd8, 0xc2,
+	0x1c, 0xc8, 0x2b, 0x73, 0x99, 0x7e, 0xeb, 0x58, 0x5b, 0xed, 0x38, 0x5a, 0x89, 0x4e, 0xa1, 0xd7,
+	0xb0, 0x26, 0xef, 0xae, 0xc7, 0x9e, 0x18, 0x9b, 0xb3, 0xab, 0x60, 0xeb, 0x58, 0xac, 0xb8, 0x0a,
+	0x52, 0x71, 0x3a, 0x57, 0x50, 0xed, 0x50, 0x69, 0x0c, 0x93, 0x3e, 0xc7, 0x82, 0xdd, 0xb2, 0xff,
+	0x61, 0x37, 0xe7, 0x0c, 0x2a, 0x96, 0xcf, 0xe8, 0x9a, 0xd0, 0x29, 0xf3, 0x10, 0x9d, 0x4e, 0x01,
+	0x45, 0x22, 0x5c, 0x52, 0x7f, 0xc0, 0xfc, 0x91, 0x1d, 0xf1, 0x3d, 0x14, 0xed, 0x15, 0x89, 0xe9,
+	0x92, 0x8e, 0x69, 0x1a, 0x00, 0x89, 0xa1, 0xce, 0x95, 0x56, 0x34, 0x26, 0x8b, 0xe7, 0x2b, 0x85,
+	0x26, 0x8e, 0xae, 0x54, 0x6e, 0xbf, 0xdc, 0x78, 0xf9, 0xcf, 0x74, 0x0b, 0x46, 0x21, 0x7f, 0xe1,
+	0x6f, 0x3e, 0x42, 0x41, 0x4f, 0x8e, 0x00, 0x0a, 0x9d, 0xee, 0x05, 0x69, 0x35, 0xab, 0x8f, 0x50,
+	0x19, 0xd6, 0x2e, 0x5b, 0xe7, 0xcd, 0xf6, 0xf9, 0xd7, 0x6a, 0x06, 0x95, 0x20, 0xff, 0xad, 0x7d,
+	0xde, 0x6a, 0x56, 0xb3, 0x51, 0xd8, 0x22, 0xe4, 0x82, 0x54, 0x73, 0x8d, 0x5f, 0x39, 0xd8, 0xb0,
+	0x9e, 0x52, 0xec, 0xe8, 0x08, 0xf2, 0x3a, 0x78, 0x91, 0xde, 0x5c, 0xfd, 0x7a, 0xfd, 0xde, 0xc9,
+	0xd0, 0x09, 0x14, 0xad, 0xad, 0xd0, 0x76, 0x02, 0x99, 0xf0, 0xdb, 0x0a, 0xa6, 0x4b, 0x58, 0x5f,
+	0x74, 0x1f, 0x72, 0x52, 0xd9, 0x96, 0xac, 0xb9, 0x82, 0xf1, 0x14, 0x4a, 0xb1, 0xd7, 0xd0, 0x4e,
+	0x12, 0x9a, 0x70, 0x61, 0x7d, 0x2b, 0xdd, 0x65, 0x96, 0xac, 0x0b, 0xe5, 0x85, 0xa3, 0x44, 0x7b,
+	0x29, 0xd3, 0x2d, 0x7b, 0xa6, 0xee, 0xdc, 0x07, 0xd1, 0xac, 0x47, 0x27, 0xdf, 0xbf, 0x8c, 0x98,
+	0x9c, 0x78, 0x3d, 0xdc, 0xe7, 0x53, 0xf7, 0x38, 0xda, 0xf3, 0x3b, 0x73, 0xe1, 0xf6, 0x27, 0x8c,
+	0xfa, 0xd2, 0x1d, 0x86, 0x6e, 0x9f, 0x87, 0xd1, 0x23, 0xe2, 0xf5, 0x6f, 0xdc, 0xe0, 0x66, 0x84,
+	0x47, 0x4c, 0xba, 0xcb, 0x0f, 0x86, 0x7e, 0x61, 0x7a, 0x05, 0xe5, 0xc3, 0xb7, 0x7f, 0x02, 0x00,
+	0x00, 0xff, 0xff, 0xb1, 0x26, 0xd8, 0xe1, 0x83, 0x06, 0x00, 0x00,
 }
 
 // Reference imports to suppress errors if they are not otherwise used.
@@ -430,241 +546,206 @@ var _ grpc.ClientConn
 // is compatible with the grpc package it is being compiled against.
 const _ = grpc.SupportPackageIsVersion4
 
-// StoreClient is the client API for Store service.
+// EnvelopeStoreClient is the client API for EnvelopeStore service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://godoc.org/google.golang.org/grpc#ClientConn.NewStream.
-type StoreClient interface {
+type EnvelopeStoreClient interface {
 	// Store an envelope
 	Store(ctx context.Context, in *StoreRequest, opts ...grpc.CallOption) (*StoreResponse, error)
-	// LoadByTxHash load an envelope by transaction hash
-	LoadByTxHash(ctx context.Context, in *TxHashRequest, opts ...grpc.CallOption) (*StoreResponse, error)
-	// LoadByID load an envelope by identifier
-	LoadByID(ctx context.Context, in *IDRequest, opts ...grpc.CallOption) (*StoreResponse, error)
-	// GetStatus returns envelope status
-	GetStatus(ctx context.Context, in *IDRequest, opts ...grpc.CallOption) (*StoreResponse, error)
+	// Load envelope by identifier
+	LoadByID(ctx context.Context, in *LoadByIDRequest, opts ...grpc.CallOption) (*StoreResponse, error)
+	// Load Envelope by transaction hash
+	LoadByTxHash(ctx context.Context, in *LoadByTxHashRequest, opts ...grpc.CallOption) (*StoreResponse, error)
 	// SetStatus set an envelope status
-	SetStatus(ctx context.Context, in *SetStatusRequest, opts ...grpc.CallOption) (*SetStatusResponse, error)
+	SetStatus(ctx context.Context, in *SetStatusRequest, opts ...grpc.CallOption) (*StatusResponse, error)
 	// LoadPending load envelopes of pending transactions
 	LoadPending(ctx context.Context, in *LoadPendingRequest, opts ...grpc.CallOption) (*LoadPendingResponse, error)
 }
 
-type storeClient struct {
+type envelopeStoreClient struct {
 	cc *grpc.ClientConn
 }
 
-func NewStoreClient(cc *grpc.ClientConn) StoreClient {
-	return &storeClient{cc}
+func NewEnvelopeStoreClient(cc *grpc.ClientConn) EnvelopeStoreClient {
+	return &envelopeStoreClient{cc}
 }
 
-func (c *storeClient) Store(ctx context.Context, in *StoreRequest, opts ...grpc.CallOption) (*StoreResponse, error) {
+func (c *envelopeStoreClient) Store(ctx context.Context, in *StoreRequest, opts ...grpc.CallOption) (*StoreResponse, error) {
 	out := new(StoreResponse)
-	err := c.cc.Invoke(ctx, "/envelopestore.Store/Store", in, out, opts...)
+	err := c.cc.Invoke(ctx, "/envelopestore.EnvelopeStore/Store", in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *storeClient) LoadByTxHash(ctx context.Context, in *TxHashRequest, opts ...grpc.CallOption) (*StoreResponse, error) {
+func (c *envelopeStoreClient) LoadByID(ctx context.Context, in *LoadByIDRequest, opts ...grpc.CallOption) (*StoreResponse, error) {
 	out := new(StoreResponse)
-	err := c.cc.Invoke(ctx, "/envelopestore.Store/LoadByTxHash", in, out, opts...)
+	err := c.cc.Invoke(ctx, "/envelopestore.EnvelopeStore/LoadByID", in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *storeClient) LoadByID(ctx context.Context, in *IDRequest, opts ...grpc.CallOption) (*StoreResponse, error) {
+func (c *envelopeStoreClient) LoadByTxHash(ctx context.Context, in *LoadByTxHashRequest, opts ...grpc.CallOption) (*StoreResponse, error) {
 	out := new(StoreResponse)
-	err := c.cc.Invoke(ctx, "/envelopestore.Store/LoadByID", in, out, opts...)
+	err := c.cc.Invoke(ctx, "/envelopestore.EnvelopeStore/LoadByTxHash", in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *storeClient) GetStatus(ctx context.Context, in *IDRequest, opts ...grpc.CallOption) (*StoreResponse, error) {
-	out := new(StoreResponse)
-	err := c.cc.Invoke(ctx, "/envelopestore.Store/GetStatus", in, out, opts...)
+func (c *envelopeStoreClient) SetStatus(ctx context.Context, in *SetStatusRequest, opts ...grpc.CallOption) (*StatusResponse, error) {
+	out := new(StatusResponse)
+	err := c.cc.Invoke(ctx, "/envelopestore.EnvelopeStore/SetStatus", in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *storeClient) SetStatus(ctx context.Context, in *SetStatusRequest, opts ...grpc.CallOption) (*SetStatusResponse, error) {
-	out := new(SetStatusResponse)
-	err := c.cc.Invoke(ctx, "/envelopestore.Store/SetStatus", in, out, opts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func (c *storeClient) LoadPending(ctx context.Context, in *LoadPendingRequest, opts ...grpc.CallOption) (*LoadPendingResponse, error) {
+func (c *envelopeStoreClient) LoadPending(ctx context.Context, in *LoadPendingRequest, opts ...grpc.CallOption) (*LoadPendingResponse, error) {
 	out := new(LoadPendingResponse)
-	err := c.cc.Invoke(ctx, "/envelopestore.Store/LoadPending", in, out, opts...)
+	err := c.cc.Invoke(ctx, "/envelopestore.EnvelopeStore/LoadPending", in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-// StoreServer is the server API for Store service.
-type StoreServer interface {
+// EnvelopeStoreServer is the server API for EnvelopeStore service.
+type EnvelopeStoreServer interface {
 	// Store an envelope
 	Store(context.Context, *StoreRequest) (*StoreResponse, error)
-	// LoadByTxHash load an envelope by transaction hash
-	LoadByTxHash(context.Context, *TxHashRequest) (*StoreResponse, error)
-	// LoadByID load an envelope by identifier
-	LoadByID(context.Context, *IDRequest) (*StoreResponse, error)
-	// GetStatus returns envelope status
-	GetStatus(context.Context, *IDRequest) (*StoreResponse, error)
+	// Load envelope by identifier
+	LoadByID(context.Context, *LoadByIDRequest) (*StoreResponse, error)
+	// Load Envelope by transaction hash
+	LoadByTxHash(context.Context, *LoadByTxHashRequest) (*StoreResponse, error)
 	// SetStatus set an envelope status
-	SetStatus(context.Context, *SetStatusRequest) (*SetStatusResponse, error)
+	SetStatus(context.Context, *SetStatusRequest) (*StatusResponse, error)
 	// LoadPending load envelopes of pending transactions
 	LoadPending(context.Context, *LoadPendingRequest) (*LoadPendingResponse, error)
 }
 
-func RegisterStoreServer(s *grpc.Server, srv StoreServer) {
-	s.RegisterService(&_Store_serviceDesc, srv)
+func RegisterEnvelopeStoreServer(s *grpc.Server, srv EnvelopeStoreServer) {
+	s.RegisterService(&_EnvelopeStore_serviceDesc, srv)
 }
 
-func _Store_Store_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+func _EnvelopeStore_Store_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(StoreRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(StoreServer).Store(ctx, in)
+		return srv.(EnvelopeStoreServer).Store(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/envelopestore.Store/Store",
+		FullMethod: "/envelopestore.EnvelopeStore/Store",
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(StoreServer).Store(ctx, req.(*StoreRequest))
+		return srv.(EnvelopeStoreServer).Store(ctx, req.(*StoreRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
 
-func _Store_LoadByTxHash_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(TxHashRequest)
+func _EnvelopeStore_LoadByID_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(LoadByIDRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(StoreServer).LoadByTxHash(ctx, in)
+		return srv.(EnvelopeStoreServer).LoadByID(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/envelopestore.Store/LoadByTxHash",
+		FullMethod: "/envelopestore.EnvelopeStore/LoadByID",
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(StoreServer).LoadByTxHash(ctx, req.(*TxHashRequest))
+		return srv.(EnvelopeStoreServer).LoadByID(ctx, req.(*LoadByIDRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
 
-func _Store_LoadByID_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(IDRequest)
+func _EnvelopeStore_LoadByTxHash_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(LoadByTxHashRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(StoreServer).LoadByID(ctx, in)
+		return srv.(EnvelopeStoreServer).LoadByTxHash(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/envelopestore.Store/LoadByID",
+		FullMethod: "/envelopestore.EnvelopeStore/LoadByTxHash",
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(StoreServer).LoadByID(ctx, req.(*IDRequest))
+		return srv.(EnvelopeStoreServer).LoadByTxHash(ctx, req.(*LoadByTxHashRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
 
-func _Store_GetStatus_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(IDRequest)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(StoreServer).GetStatus(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: "/envelopestore.Store/GetStatus",
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(StoreServer).GetStatus(ctx, req.(*IDRequest))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
-func _Store_SetStatus_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+func _EnvelopeStore_SetStatus_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(SetStatusRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(StoreServer).SetStatus(ctx, in)
+		return srv.(EnvelopeStoreServer).SetStatus(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/envelopestore.Store/SetStatus",
+		FullMethod: "/envelopestore.EnvelopeStore/SetStatus",
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(StoreServer).SetStatus(ctx, req.(*SetStatusRequest))
+		return srv.(EnvelopeStoreServer).SetStatus(ctx, req.(*SetStatusRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
 
-func _Store_LoadPending_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+func _EnvelopeStore_LoadPending_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(LoadPendingRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(StoreServer).LoadPending(ctx, in)
+		return srv.(EnvelopeStoreServer).LoadPending(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/envelopestore.Store/LoadPending",
+		FullMethod: "/envelopestore.EnvelopeStore/LoadPending",
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(StoreServer).LoadPending(ctx, req.(*LoadPendingRequest))
+		return srv.(EnvelopeStoreServer).LoadPending(ctx, req.(*LoadPendingRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
 
-var _Store_serviceDesc = grpc.ServiceDesc{
-	ServiceName: "envelopestore.Store",
-	HandlerType: (*StoreServer)(nil),
+var _EnvelopeStore_serviceDesc = grpc.ServiceDesc{
+	ServiceName: "envelopestore.EnvelopeStore",
+	HandlerType: (*EnvelopeStoreServer)(nil),
 	Methods: []grpc.MethodDesc{
 		{
 			MethodName: "Store",
-			Handler:    _Store_Store_Handler,
-		},
-		{
-			MethodName: "LoadByTxHash",
-			Handler:    _Store_LoadByTxHash_Handler,
+			Handler:    _EnvelopeStore_Store_Handler,
 		},
 		{
 			MethodName: "LoadByID",
-			Handler:    _Store_LoadByID_Handler,
+			Handler:    _EnvelopeStore_LoadByID_Handler,
 		},
 		{
-			MethodName: "GetStatus",
-			Handler:    _Store_GetStatus_Handler,
+			MethodName: "LoadByTxHash",
+			Handler:    _EnvelopeStore_LoadByTxHash_Handler,
 		},
 		{
 			MethodName: "SetStatus",
-			Handler:    _Store_SetStatus_Handler,
+			Handler:    _EnvelopeStore_SetStatus_Handler,
 		},
 		{
 			MethodName: "LoadPending",
-			Handler:    _Store_LoadPending_Handler,
+			Handler:    _EnvelopeStore_LoadPending_Handler,
 		},
 	},
 	Streams:  []grpc.StreamDesc{},